arreglos varios
