+++ .env.copy
@@ -1,13 +1,17 @@
-# MÓDULO DE VARIABLES DE ENTORNO MUESTRA
+# MÓDULO DE VARIABLES DE ENTORNO (COPIA)
+#(Copiar el codigo en un .env vuestro y rellenar datos)
 
 # Configuración de la base de datos MySQL
-DB_HOST=localhost
-DB_USER=
-DB_PASSWORD=
-DB_NAME=app_notas
+MYSQL_HOST=localhost
+MYSQL_USER=
+MYSQL_PASS=
+MYSQL_DB=app_notas
 
 # Configuración del servidor
 PORT=3000
 
-# Clave secreta para JWT (genera una clave segura)
+# Clave secreta para JWT (genera una clave segura)!!
 JWT_SECRET=
+
+# Otras configuraciones!!
+LOG_LEVEL=debug
