+++ package.json
@@ -1,21 +1,21 @@
+
 {
   "type": "module",
-  "name": "bloc-de-notas-",
+  "name": "bloc-de-notas",
   "version": "1.0.0",
-  "description": "",
-  "main": "main.js",
+  "description": "Aplicación de notas de texto",
+  "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon main.js"
+    "dev": "nodemon index.js"
   },
-  "keywords": [],
+  "keywords": ["notas", "texto", "categorización"],
   "author": "Grupo C, App de notas",
   "license": "ISC",
   "dependencies": {
-    "-": "^0.0.1",
     "bcrypt": "^5.1.1",
+    "bcryptjs": "^2.4.3",
     "cors": "^2.8.5",
-    "d": "^1.0.1",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "express-fileupload": "^1.4.3",
@@ -27,9 +27,9 @@
     "nodemon": "^3.0.2",
     "prettier": "^3.1.1",
     "randomstring": "^1.3.0",
-    "sharp": "^0.33.0"
+    "sharp": "^0.33.1"
   },
   "devDependencies": {
-    "eslint": "^8.55.0"
+    "eslint": "^8.56.0"
   }
 }
