--- src/middleware/JasonWebToken/authConfig.js
@@ -1,6 +1,0 @@
-//MÓDULO DE CONFIGURACIÓN DE AUTENTIFICACIÓN CON JWT
-
-module.exports = {
-  JWT_SECRET: '',  // Clave secreta para firmar el token
-  JWT_EXPIRATION: '1d'// Tiempo de expiración del token
-};
